--- a/core/src/main/java/org/jruby/util/RubyDateFormatter.java
+++ b/core/src/main/java/org/jruby/util/RubyDateFormatter.java
@@ -1,4 +1,5 @@
-/***** BEGIN LICENSE BLOCK *****
+/*
+ **** BEGIN LICENSE BLOCK *****
  * Version: EPL 2.0/GPL 2.0/LGPL 2.1
  *
  * The contents of this file are subject to the Eclipse Public
@@ -31,18 +32,16 @@
 
 package org.jruby.util;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.Reader;
 import java.text.DateFormat;
 import java.text.DateFormatSymbols;
 import java.text.ParsePosition;
 import java.util.Calendar;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
+import java.util.Map;
 
 import jnr.constants.platform.Errno;
 import org.jcodings.Encoding;
@@ -58,9 +57,15 @@ import org.jruby.RubyTime;
 import org.jruby.lexer.StrftimeLexer;
 import org.jruby.runtime.ThreadContext;
 
+import static org.jruby.ext.stringio.StringIO.NEWLINE;
+import static org.jruby.util.CommonByteLists.*;
 import static org.jruby.util.RubyDateFormatter.FieldType.*;
 
 public class RubyDateFormatter {
+    private static final ByteList AM = new ByteList(new byte[] {'a', 'm'});
+    private static final ByteList PM = new ByteList(new byte[] {'p', 'm'});
+    private static final ByteList CAPITAL_AM = new ByteList(new byte[] {'A', 'M'});
+    private static final ByteList CAPITAL_PM = new ByteList(new byte[] {'P', 'M'});
 
     private static final String[] FORMAT_MONTHS;
     private static final String[] FORMAT_SHORT_MONTHS;
@@ -193,7 +198,7 @@ public class RubyDateFormatter {
             this.data = data;
         }
 
-        public static Token str(String str) {
+        public static Token str(ByteList str) {
             return new Token(Format.FORMAT_STRING, str);
         }
 
@@ -241,7 +246,7 @@ public class RubyDateFormatter {
     public RubyDateFormatter(ThreadContext context) {
         super();
         this.runtime = context.runtime;
-        lexer = new StrftimeLexer((Reader) null);
+        lexer = new StrftimeLexer();
     }
 
     private static void addToPattern(List<Token> compiledPattern, String str) {
@@ -250,7 +255,7 @@ public class RubyDateFormatter {
             if (('A' <= c && c <= 'Z') || ('a' <= c && c <= 'z')) {
                 compiledPattern.add(Token.format(c));
             } else {
-                compiledPattern.add(Token.str(Character.toString(c)));
+                compiledPattern.add(Token.str(new ByteList(new byte[] { (byte) c }))); // FIXME broken patterns hsould pass in bytelists
             }
         }
     }
@@ -270,13 +275,10 @@ public class RubyDateFormatter {
             compiledPattern.add(new Token(Format.FORMAT_ENCODING, enc));
         }
 
-        ByteArrayInputStream in = new ByteArrayInputStream(pattern.getUnsafeBytes(), pattern.getBegin(), pattern.getRealSize());
-        Reader reader = new InputStreamReader(in, runtime.getEncodingService().charsetForEncoding(pattern.getEncoding()));
-        lexer.yyreset(reader);
+        lexer.reset(pattern);
 
         Token token;
         RubyTimeOutputFormatter formatter = null;
-        try {
             while ((token = lexer.yylex()) != null) {
                 if (token.format == Format.FORMAT_OUTPUT) {
                     formatter = (RubyTimeOutputFormatter) token.data;
@@ -294,8 +296,8 @@ public class RubyDateFormatter {
                         switch (c) {
                             default:
                                 // force most formats to use spaces for padding unless otherwise specified
-                                if (formatter.flags == "" || formatter.flags == null) {
-                                    compiledPattern.add(new Token(Format.FORMAT_OUTPUT, new RubyTimeOutputFormatter("_", formatter.width)));
+                                if (formatter.flags == ByteList.EMPTY_BYTELIST || formatter.flags == null) {
+                                    compiledPattern.add(new Token(Format.FORMAT_OUTPUT, new RubyTimeOutputFormatter(UNDERSCORE, formatter.width)));
                                     break;
                                 }
                                 // fall through
@@ -318,13 +320,13 @@ public class RubyDateFormatter {
                         addToPattern(compiledPattern, "Y-m-d");
                         break;
                     case 'n':
-                        compiledPattern.add(Token.str("\n"));
+                        compiledPattern.add(Token.str(NEWLINE));
                         break;
                     case 'Q':
                         if (dateLibrary) {
                             compiledPattern.add(new Token(Format.FORMAT_MICROSEC_EPOCH));
                         } else {
-                            compiledPattern.add(Token.str("%Q"));
+                            compiledPattern.add(Token.str(PERCENT_Q));
                         }
                         break;
                     case 'R':
@@ -338,12 +340,12 @@ public class RubyDateFormatter {
                         addToPattern(compiledPattern, "H:M:S");
                         break;
                     case 't':
-                        compiledPattern.add(Token.str("\t"));
+                        compiledPattern.add(Token.str(TAB));
                         break;
                     case 'v':
                         addToPattern(compiledPattern, "e-");
                         if (!dateLibrary)
-                            compiledPattern.add(Token.formatter(new RubyTimeOutputFormatter("^", 0)));
+                            compiledPattern.add(Token.formatter(new RubyTimeOutputFormatter(CARET, 0)));
                         addToPattern(compiledPattern, "b-Y");
                         break;
                     case 'Z':
@@ -356,7 +358,7 @@ public class RubyDateFormatter {
                         break;
                     case '+':
                         if (!dateLibrary) {
-                            compiledPattern.add(Token.str("%+"));
+                            compiledPattern.add(Token.str(PERCENT_PLUS));
                             break;
                         }
                         addToPattern(compiledPattern, "a b e H:M:S ");
@@ -374,9 +376,6 @@ public class RubyDateFormatter {
             if (formatter != null) {
                 compiledPattern.add(new Token(Format.FORMAT_OUTPUT, formatter));
             }
-        } catch (IOException e) {
-            throw new AssertionError(e); // IOException never happens
-        }
 
         return compiledPattern;
     }
@@ -400,7 +399,8 @@ public class RubyDateFormatter {
 
     /** Convenience method when using no pattern caching */
     public RubyString compileAndFormat(RubyString pattern, boolean dateLibrary, DateTime dt, long nsec, RubyNumeric sub_millis) {
-        RubyString out = format(compilePattern(pattern, dateLibrary), dt, nsec, sub_millis);
+        List<Token>  compiled = compilePattern(pattern, dateLibrary);
+        RubyString out = format(compiled, dt, nsec, sub_millis);
         out.setEncoding(pattern.getEncoding());
         if (pattern.isTaint()) out.setTaint(true);
         return out;
@@ -412,39 +412,39 @@ public class RubyDateFormatter {
 
     private ByteList formatToByteList(List<Token> compiledPattern, DateTime dt, long nsec, RubyNumeric sub_millis) {
         RubyTimeOutputFormatter formatter = RubyTimeOutputFormatter.DEFAULT_FORMATTER;
-        final ByteList toAppendTo = new ByteList(24);
+        final ByteList output = new ByteList(24);
 
         for (Token token: compiledPattern) {
-            CharSequence output = null;
+            CharSequence data = null;
             long value = 0;
             FieldType type = TEXT;
             Format format = token.getFormat();
 
             switch (format) {
                 case FORMAT_ENCODING:
-                    toAppendTo.setEncoding((Encoding) token.getData());
+                    output.setEncoding((Encoding) token.getData());
                     continue; // go to next token
                 case FORMAT_OUTPUT:
                     formatter = (RubyTimeOutputFormatter) token.getData();
                     continue; // go to next token
                 case FORMAT_STRING:
-                    output = token.getData().toString();
+                    data = (ByteList) token.getData();
                     break;
                 case FORMAT_WEEK_LONG:
                     // This is GROSS, but Java API's aren't ISO 8601 compliant at all
                     int v = (dt.getDayOfWeek() + 1) % 8;
-                    output = FORMAT_WEEKDAYS[v == 0 ? 1 : v];
+                    data = FORMAT_WEEKDAYS[v == 0 ? 1 : v];
                     break;
                 case FORMAT_WEEK_SHORT:
                     // This is GROSS, but Java API's aren't ISO 8601 compliant at all
                     v = (dt.getDayOfWeek() + 1) % 8;
-                    output = FORMAT_SHORT_WEEKDAYS[v == 0 ? 1 : v];
+                    data = FORMAT_SHORT_WEEKDAYS[v == 0 ? 1 : v];
                     break;
                 case FORMAT_MONTH_LONG:
-                    output = FORMAT_MONTHS[dt.getMonthOfYear() - 1];
+                    data = FORMAT_MONTHS[dt.getMonthOfYear() - 1];
                     break;
                 case FORMAT_MONTH_SHORT:
-                    output = FORMAT_SHORT_MONTHS[dt.getMonthOfYear() - 1];
+                    data = FORMAT_SHORT_MONTHS[dt.getMonthOfYear() - 1];
                     break;
                 case FORMAT_DAY:
                     type = NUMERIC2;
@@ -486,10 +486,10 @@ public class RubyDateFormatter {
                     value = dt.getMonthOfYear();
                     break;
                 case FORMAT_MERIDIAN:
-                    output = dt.getHourOfDay() < 12 ? "AM" : "PM";
+                    data = dt.getHourOfDay() < 12 ? CAPITAL_AM : CAPITAL_PM;
                     break;
                 case FORMAT_MERIDIAN_LOWER_CASE:
-                    output = dt.getHourOfDay() < 12 ? "am" : "pm";
+                    data = dt.getHourOfDay() < 12 ? AM : PM;
                     break;
                 case FORMAT_SECONDS:
                     type = NUMERIC2;
@@ -523,10 +523,11 @@ public class RubyDateFormatter {
                     // custom logic because this is so weird
                     value = dt.getZone().getOffset(dt.getMillis()) / 1000;
                     int colons = (Integer) token.getData();
-                    output = formatZone(colons, (int) value, formatter);
+                    data = formatZone(colons, (int) value, formatter);
                     break;
                 case FORMAT_ZONE_ID:
-                    output = RubyTime.getRubyTimeZoneName(runtime, dt);
+                    // Should be safe to assume all time zone labels will be ASCII 7bit.
+                    data = RubyTime.getRubyTimeZoneName(runtime, dt);
                     break;
                 case FORMAT_CENTURY:
                     type = NUMERIC;
@@ -545,24 +546,24 @@ public class RubyDateFormatter {
                     int defaultWidth = (format == Format.FORMAT_NANOSEC) ? 9 : 3;
                     int width = formatter.getWidth(defaultWidth);
 
-                    output = RubyTimeOutputFormatter.formatNumber(dt.getMillisOfSecond(), 3, '0');
+                    data = new ByteList();
+                    RubyTimeOutputFormatter.formatNumber((ByteList) data, dt.getMillisOfSecond(), 3, '0');
                     if (width > 3) {
-                        StringBuilder buff = new StringBuilder(output.length() + 6).append(output);
                         if (sub_millis == null) { // Time
-                            buff.append(RubyTimeOutputFormatter.formatNumber(nsec, 6, '0'));
+                            RubyTimeOutputFormatter.formatNumber((ByteList) data, nsec, 6, '0');
                         } else { // Date, DateTime
-                            formatSubMillisGt3(runtime, buff, width, sub_millis);
+                            formatSubMillisGt3(runtime, (ByteList) data, width, sub_millis);
                         }
-                        output = buff;
                     }
 
-                    if (width < output.length()) {
-                        output = output.subSequence(0, width);
+                    if (width < data.length()) {
+                        data = ((ByteList) data).makeShared(0, width);
                     } else {
-                        StringBuilder buff = new StringBuilder(width).append(output);
-                        // Not enough precision, fill with 0
-                        while (buff.length() < width) buff.append('0');
-                        output = buff;
+                        int padLength = width - data.length();
+                        // FIXME: length - width fill of 0's can be pre-calc'd
+                        for (int i = 0; i < padLength; i++) {  // Not enough precision, fill with 0
+                            ((ByteList) data).append('0');
+                        }
                     }
                     formatter = RubyTimeOutputFormatter.DEFAULT_FORMATTER; // no more formatting
                     break;
@@ -583,23 +584,24 @@ public class RubyDateFormatter {
                     throw new Error("FORMAT_SPECIAL is a special token only for the lexer.");
             }
 
-            final String formatted;
             try {
-                formatted = formatter.format(output, value, type);
+                if (data == null) {
+                    formatter.format(output, value, type);
+                } else {
+                    formatter.format(output, data, type);
+                }
             } catch (IndexOutOfBoundsException ioobe) {
                 throw runtime.newErrnoFromErrno(Errno.ERANGE, "strftime");
             }
 
             // reset formatter
             formatter = RubyTimeOutputFormatter.DEFAULT_FORMATTER;
-
-            toAppendTo.append(formatted.getBytes(runtime.getEncodingService().charsetForEncoding(toAppendTo.getEncoding())));
         }
 
-        return toAppendTo;
+        return output;
     }
 
-    private static void formatSubMillisGt3(final Ruby runtime, final StringBuilder buff,
+    private static void formatSubMillisGt3(final Ruby runtime, final ByteList buff,
                                            final int width, RubyNumeric sub_millis) {
         final int prec = width - 3;
         final ThreadContext context = runtime.getCurrentContext();
@@ -608,7 +610,7 @@ public class RubyDateFormatter {
                 convertToInteger().op_mul(context, power);
         truncated = (RubyNumeric) truncated.idiv(context, sub_millis.denominator(context));
         long decimals = truncated.convertToInteger().getLongValue();
-        buff.append(RubyTimeOutputFormatter.formatNumber(decimals, prec, '0'));
+        RubyTimeOutputFormatter.formatNumber(buff, decimals, prec, '0');
     }
 
     /**
@@ -638,7 +640,7 @@ public class RubyDateFormatter {
         return value;
     }
 
-    private static StringBuilder formatZone(int colons, int value, RubyTimeOutputFormatter formatter) {
+    private static ByteList formatZone(int colons, int value, RubyTimeOutputFormatter formatter) {
         int seconds = Math.abs(value);
         int hours = seconds / 3600;
         seconds %= 3600;
@@ -649,31 +651,36 @@ public class RubyDateFormatter {
             hours = -hours;
         }
 
-        CharSequence mm = RubyTimeOutputFormatter.formatNumber(minutes, 2, '0');
-        CharSequence ss = RubyTimeOutputFormatter.formatNumber(seconds, 2, '0');
-
         char padder = formatter.getPadder('0');
         int defaultWidth = -1;
-        CharSequence after = null;
+        ByteList after = new ByteList(12);
 
         switch (colons) {
             case 0: // %z -> +hhmm
                 defaultWidth = 5;
-                after = mm;
+                RubyTimeOutputFormatter.formatNumber(after, minutes, 2, '0');
                 break;
             case 1: // %:z -> +hh:mm
                 defaultWidth = 6;
-                after = new StringBuilder(mm.length() + 1).append(':').append(mm);
+                after.append(':');
+                RubyTimeOutputFormatter.formatNumber(after, minutes, 2, '0');
                 break;
             case 2: // %::z -> +hh:mm:ss
                 defaultWidth = 9;
-                after = new StringBuilder(mm.length() + ss.length() + 2).append(':').append(mm).append(':').append(ss);
+                after.append(':');
+                RubyTimeOutputFormatter.formatNumber(after, minutes, 2, '0');
+                after.append(':');
+                RubyTimeOutputFormatter.formatNumber(after, seconds, 2, '0');
                 break;
             case 3: // %:::z -> +hh[:mm[:ss]]
-                StringBuilder sb = new StringBuilder(mm.length() + ss.length() + 2);
-                if (minutes != 0 || seconds != 0) sb.append(':').append(mm);
-                if (seconds != 0) sb.append(':').append(ss);
-                after = sb;
+                if (minutes != 0 || seconds != 0) {
+                    after.append(':');
+                    RubyTimeOutputFormatter.formatNumber(after, minutes, 2, '0');
+                }
+                if (seconds != 0) {
+                    after.append(':');
+                    RubyTimeOutputFormatter.formatNumber(after, seconds, 2, '0');
+                }
                 defaultWidth = after.length() + 3;
                 break;
         }
@@ -683,15 +690,11 @@ public class RubyDateFormatter {
         if (width < minWidth) {
             width = minWidth;
         }
+        ByteList result = new ByteList(width);
         width -= after.length();
-        StringBuilder before = RubyTimeOutputFormatter.formatSignedNumber(hours, width, padder);
-
-        if (value < 0 && hours == 0) { // the formatter could not handle this case
-            for (int i=0; i<before.length(); i++) { // replace('+', '-')
-                if (before.charAt(i) == '+') before.setCharAt(i, '-');
-            }
-        }
-        return new StringBuilder(before.length() + after.length()).append(before).append(after); // before + after
+        RubyTimeOutputFormatter.formatSignedNumber(result, hours, value, width, padder);
+        result.append(after); // before + after
+        return result;
     }
 
     /**