… core/src/main/java/org/jruby/lexer/StrftimeLexer.flex
… core/src/main/java/org/jruby/util/RubyTimeOutputFormatter.java